backend/requirements.txt
# Web framework
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic-settings==2.9.1

# Database
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
alembic==1.15.2

# Auth
authlib==1.5.2
httpx==0.28.1
pyjwt[crypto]==2.10.1

# Storage
boto3==1.38.0

# AI
anthropic==0.52.0

# Push notifications
pywebpush==2.0.1

# Scheduling
apscheduler==3.11.0

# Utilities
python-dotenv==1.1.0
python-multipart==0.0.20
email-validator==2.2.0
itsdangerous==2.2.0

# Testing
pytest==8.3.5
pytest-asyncio==0.26.0
